Rename user defined publisher file
